VM scripts should now work with with Debian 7 and Centos 6

+++ docs/install/VM/setup_nginx.sh
@@ -2,11 +2,23 @@
 
 set -e -u -x
 
-NGINX_CONF=/etc/nginx/sites-available/omero-web
 $OMERO_PREFIX/bin/omero web config nginx --system --http "$OMERO_WEB_PORT" > nginx.tmp
-sudo cp nginx.tmp ${NGINX_CONF}
-rm nginx.tmp
 
-sudo rm /etc/nginx/sites-enabled/default
-sudo ln -s ${NGINX_CONF} /etc/nginx/sites-enabled/
+if [ -d /etc/nginx/sites-available/ -a -d /etc/nginx/sites-enabled/ ]; then
+    # Debian
+    NGINX_CONF=/etc/nginx/sites-available/omero-web
+    sudo cp nginx.tmp ${NGINX_CONF}
+    sudo rm /etc/nginx/sites-enabled/default
+    sudo ln -s ${NGINX_CONF} /etc/nginx/sites-enabled/
+elif [ -d /etc/nginx/conf.d/ ]; then
+    # Redhat
+    NGINX_CONF=/etc/nginx/conf.d/omero-web.conf
+    sudo mv /etc/nginx/conf.d/default.conf /etc/nginx/conf.d/default.disabled
+    sudo cp nginx.tmp ${NGINX_CONF}
+else
+    echo "ERROR: Unable to find nginx configuration directory"
+    exit 2
+fi
+
+rm nginx.tmp
